+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - [9258](https://github.com/vegaprotocol/vega/issues/9258) - Change HTTP endpoint from `/volume-discount-stats` to `/volume-discount-program/stats`.
 - [9258](https://github.com/vegaprotocol/vega/issues/9258) - Change HTTP endpoint from `/referral-sets/stats/{id}` to `/referral-sets/{id}/stats`.
 - [9719](https://github.com/vegaprotocol/vega/issues/9719) - Remove unnecessary fields from referral and volume discount program proposals.
+- [9733](https://github.com/vegaprotocol/vega/issues/9733) - Making `set_id` optional in `referral set stats` endpoint
 
 ### 🗑️ Deprecation
 
@@ -296,6 +297,7 @@
 - [9727](https://github.com/vegaprotocol/vega/issues/9727) - Initialize teams earlier to avoid panic.
 - [9746](https://github.com/vegaprotocol/vega/issues/9746) - Fix handling of LP fees reward 
 - [9747](https://github.com/vegaprotocol/vega/issues/9747) - Return correct destination type
+- [9541](https://github.com/vegaprotocol/vega/issues/9731) - Add filtering for party to the referral fees API.
 
 ## 0.72.1